requirements.txt
python-dotenv
streamlit
langchain_community
langchain_objectbox
langchain_openai 
langchain_core
langchain
bs4
langchainhub
langchain_groq
langchain_objectbox
PyPDF2
pypdf